分离webwxbatchgetcontact，添加截图
