+++ README.md
@@ -68,6 +68,10 @@ request:
 
 [如果您觉得这些东西对您有用，请支持自由软件基金会](https://my.fsf.org/donate/?pk_campaign=2015-2016-fundraiser-banner-gnu&pk_kwd=donate)
 
+## 截图
+
+![截图](https://raw.githubusercontent.com/HalfdogStudio/wechat-user-bot/refactor/screenshot/0.1.3.png)
+
 ## ChangeLog
 
 ### 2016.2.9
