+++ lib/logger/logger.js
@@ -5,6 +5,7 @@ var fs = require('fs');
 var querystring = require('querystring');
 var crypto = require('crypto');
 var path = require('path');
+var webwxbatchgetcontact = require('../webwx.js').webwxbatchgetcontact;
 
 var MSGTYPE_TEXT = require('../global.js').MSGTYPE_TEXT;
 var MSGTYPE_IMAGE = require('../global.js').MSGTYPE_IMAGE;
@@ -26,17 +27,20 @@ function wechatLogger(obj) {
   }
 }
 
-function logImageMessage(o, obj) {
-  //debug("in webwxsync someone call me:" + inspect(o));
-  // 查询用户名昵称
-  var imgUrl = `https://wx.qq.com/cgi-bin/mmwebwx-bin/webwxgetmsgimg?&MsgID=${o.MsgId}&skey=${querystring.escape(obj.BaseRequest.Skey)}`;
+
+function webwxgetmsgimg(msgId, obj, imgPath){
+  var imgUrl = `https://wx.qq.com/cgi-bin/mmwebwx-bin/webwxgetmsgimg?&MsgID=${msgId}&skey=${querystring.escape(obj.BaseRequest.Skey)}`;
   // 保存图片到文件
-  var imgPath = path.join(process.cwd(), 'data/pic', crypto.createHash('md5').update(crypto.randomBytes(10)).digest('hex') + '.jpg');
   try {
     request.get(imgUrl, {jar: true}).pipe(fs.createWriteStream(imgPath));
   } catch (e){
     console.error('下载图像资源失败:', e);
   }
+}
+
+function logImageMessage(o, obj) {
+  var imgPath = path.join(process.cwd(), 'data/pic', crypto.createHash('md5').update(crypto.randomBytes(10)).digest('hex') + '.jpg');
+  webwxgetmsgimg(o.MsgId, obj, imgPath);
   if (o.FromUserName.startsWith("@@")) {
     logGroupImageMsg(o, obj, imgPath);
   } else {
@@ -80,7 +84,7 @@ function handlePrivate(username, replyContent, obj) {
   var p = new Promise((resolve, reject)=>{
     if (obj.memberList.findIndex(m=>m['UserName']==username) < 0) { 
       // memberList中不存在
-      var contactP = _requestUserInfo(username, obj);
+      var contactP = webwxbatchgetcontact(username, obj);
     } else {
       var contactP = Promise.resolve(obj);
     }
@@ -91,48 +95,9 @@ function handlePrivate(username, replyContent, obj) {
       var m = obj.memberList.find(m=>m.UserName==username);
       resolve("[" + m.NickName + "说]" + replyContent);
     }
-
-    function _requestUserInfo(username, obj) {
-      return new Promise((resolve, reject)=>{
-        var postData = {
-          BaseRequest: obj.BaseRequest,
-          Count: 1,
-          List: [
-            {
-              UserName: username,
-              EncryChatRoomId: "",
-            }
-          ]
-        };
-        // console.log("为啥Promise里看不到运行情况")
-        request.post(`https://wx.qq.com/cgi-bin/mmwebwx-bin/webwxbatchgetcontact`,
-                     {
-                       qs: {
-                         type: 'ex',
-                         r: Date.now(),
-                       },
-                       body: postData,
-                       json: true,
-                       jar: true,
-                     },
-                     (error, response, body)=> {
-                       //console.log(body);
-                       if (error) {
-                         reject(error)
-                       }
-                       if (body.BaseResponse.Ret != 0) {
-                         reject(body.BaseResponse.ErrMsg);
-                       }
-                       var user = body.ContactList[0]
-                       obj.memberList.push(user);
-                       resolve(obj);
-                     });
-      });
-    }
   });
   return p;
 }
-
 
 function logGroupTextMsg(o, obj) {
   var p = handleGroup(o.FromUserName, o.Content, obj);
@@ -149,7 +114,7 @@ function handleGroup(groupUserName, replyContent, obj) {
     }
     // 查看是否缓存中有
     if (!(groupUserName in obj.groupContact)) {
-      var contactP = _requestGroupInfo(groupUserName, obj)
+      var contactP = webwxbatchgetcontact(groupUserName, obj)
     } else {
       var contactP = Promise.resolve(obj);
     }
@@ -162,51 +127,10 @@ function handleGroup(groupUserName, replyContent, obj) {
       resolve("[" + groupRealName + "]" + m.NickName + replyContent.replace(fromUserName, '').replace("<br/>", ""));
     }
 
-    function _requestGroupInfo(groupUserName, obj) {
-      return new Promise((resolve, reject)=>{
-        var postData = {
-          BaseRequest: obj.BaseRequest,
-          Count: 1,
-          List: [
-            {
-              UserName: groupUserName,
-              EncryChatRoomId: "",
-            }
-          ]
-        };
-        request.post(`https://wx.qq.com/cgi-bin/mmwebwx-bin/webwxbatchgetcontact`,
-                     {
-                       qs: {
-                         type: 'ex',
-                         r: Date.now(),
-                       },
-                       body: postData,
-                       json: true,
-                       jar: true,
-                     },
-                     (error, response, body)=> {
-                       //console.log(body);
-                       if (error) {
-                         reject(error)
-                       }
-                       if (body.BaseResponse.Ret != 0) {
-                         reject(body.BaseResponse.ErrMsg);
-                       }
-                       var group = body.ContactList[0]
-                       var groupRealName = group.NickName;
-                       var memberList = group.MemberList;
-                       obj.groupContact[groupUserName] = {
-                         memberList: memberList,
-                         nickName: groupRealName, 
-                       };
-                       resolve(obj);
-                     });    // request
-      });   //promise
-    }
   });
   return p;
-
 }
+
 
 function logNotImplementMsg(o) {
   console.error("未实现消息类型：" + o.MsgType);
