+++ lib/webwx.js
@@ -395,6 +395,53 @@ function webwxsync(filters, mappers) {
   }
 }
 
+function webwxbatchgetcontact(username, obj) {
+  return new Promise((resolve, reject)=>{
+    var postData = {
+      BaseRequest: obj.BaseRequest,
+      Count: 1,
+      List: [
+        {
+          UserName: username,
+          EncryChatRoomId: "",
+        }
+      ]
+    };
+    // console.log("为啥Promise里看不到运行情况")
+    request.post(`https://wx.qq.com/cgi-bin/mmwebwx-bin/webwxbatchgetcontact`,
+                 {
+                   qs: {
+                     type: 'ex',
+                     r: Date.now(),
+                   },
+                   body: postData,
+                   json: true,
+                   jar: true,
+                 },
+                 (error, response, body)=> {
+                   //console.log(body);
+                   if (error) {
+                     reject(error)
+                   }
+                   if (body.BaseResponse.Ret != 0) {
+                     reject(body.BaseResponse.ErrMsg);
+                   }
+                   if (!username.startsWith('@@')) {    // 个人
+                     var user = body.ContactList[0]
+                     obj.memberList.push(user);
+                   } else { // 群组
+                     var group = body.ContactList[0]
+                     var groupRealName = group.NickName;
+                     var memberList = group.MemberList;
+                     obj.groupContact[username] = {
+                       memberList: memberList,
+                       nickName: groupRealName, 
+                     };
+                   }
+                   resolve(obj);
+                 });
+  });
+}
 
 function sessionStop(code, signal) {
   console.log('结束会话');
@@ -413,3 +460,4 @@ module.exports.getContact = getContact;
 module.exports.synccheck = synccheck;
 module.exports.webwxsync = webwxsync;
 module.exports.webwxsendmsg = webwxsendmsg;
+module.exports.webwxbatchgetcontact = webwxbatchgetcontact;
